fixing navbar add Protected
