+++ redux/reducers/index.js
@@ -1,5 +1,4 @@
 import { combineReducers } from "@reduxjs/toolkit";
-import auth from "./auth";
 import flights from "./flight";
 import auth from "./auth";
